Added Piranha in Pipe.
Changed some DrawOrder stuff too so all revealed items will appear beneath their respective block/pipe.

+++ MarioClone/GameObjects/IDraw.cs
@@ -10,7 +10,7 @@ namespace MarioClone.GameObjects
 {
     public interface IDraw 
     {
-        int DrawOrder { get; }
+        float DrawOrder { get; }
         bool Visible { get; }
 
         void Draw(SpriteBatch spriteBatch, GameTime gameTime);
